calendar_app/requirements.txt
Django==1.10.5
djangorestframework==3.5.3
google-api-python-client==1.6.1
httplib2==0.9.2
iso8601==0.1.11
oauth2client==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pytz==2016.10
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
